requirements.txt
pandas==1.4.2
numpy==1.26.4
boto3==1.34.102
fastapi==0.111.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn[standard]==0.15.0
